+++ angular/index.run.js
@@ -1,0 +1,4 @@
+import {RoutesRun} from './run/routes.run';
+
+angular.module('app.run')
+    .run(RoutesRun);