+++ angular/run/routes.run.js
@@ -1,0 +1,15 @@
+export function RoutesRun($rootScope, $state, $auth) {
+    'ngInject';
+
+    $rootScope.$on("$stateChangeStart", function(event, toState) {
+
+        if (toState.data && toState.data.auth) {
+            /*Cancel going to the authenticated state and go back to landing*/
+            if (!$auth.isAuthenticated()) {
+                event.preventDefault();
+                return $state.go('app.landing');
+            }
+        }
+
+    });
+}